requirements.txt
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
html5lib==1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mimerender==0.6.0
Pillow==7.2.0
PyPDF2==1.26.0
python-dotenv==0.14.0
python-mimeparse==1.6.0
pytz==2020.1
qrcode==6.1
qrplatba==0.3.4
reportlab==3.5.44
requests==2.24.0
six==1.15.0
urllib3==1.25.9
webencodings==0.5.1
Werkzeug==1.0.1
xhtml2pdf==0.2.4
